+++ runtime/parachain/Cargo.toml
@@ -77,6 +77,18 @@ parachain-info = { git = 'https://github.com/paritytech/cumulus', branch = 'polk
 orml-traits = { git = "https://github.com/t3rn/open-runtime-module-library", branch = "polkadot-v0.9.17", default-features = false }
 orml-tokens = { git = "https://github.com/t3rn/open-runtime-module-library", branch = "polkadot-v0.9.17", default-features = false }
 
+# t3rn Deps
+pallet-randomness-collective-flip = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "polkadot-v0.9.17" }
+t3rn-primitives = { default-features = false, path = "../../primitives" }
+t3rn-protocol = { default-features = false, path = "../../protocol" }
+
+# t3rn pallets
+pallet-xdns = {path = "../../pallets/xdns", default-features = false }
+pallet-mfv = {path = "../../pallets/multi-finality-verifier", package = "pallet-multi-finality-verifier", default-features = false }
+pallet-contracts-registry = { path = "../../pallets/contracts-registry", default-features = false }
+pallet-circuit-portal = { path = "../../pallets/circuit-portal", package = "pallet-circuit-portal", default-features = false }
+pallet-circuit = {path = "../../pallets/circuit", package = "pallet-circuit", default-features = false }
+
 [features]
 default = [
 	"std",
@@ -126,8 +138,19 @@ std = [
 	"xcm-builder/std",
 	"xcm-executor/std",
 	"xcm/std",
+	# ORML
 	"orml-tokens/std",
 	"orml-traits/std",
+	# t3rn deps
+	"pallet-randomness-collective-flip/std",
+	"t3rn-primitives/std",
+	"t3rn-protocol/std",
+	# t3rn pallets
+	"pallet-xdns/std",
+	"pallet-mfv/std",
+	"pallet-contracts-registry/std",
+	"pallet-circuit-portal/std",
+	"pallet-circuit/std",
 ]
 
 runtime-benchmarks = [
